third_party/WebKit/LayoutTests/accessibility/focusable-div-expected.txt

A
B
C
Link
Initial text before linkLink
List item
Initial text before list
List item

This test makes sure that a generic focusable div can get accessibility focus and gets its accessible text from contents..

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement == link is true
PASS axLink.name is "A"
PASS document.activeElement == div is true
PASS axDiv.name is "B"
PASS document.activeElement == div2 is true
PASS axDiv2.name is "C"
PASS document.activeElement == div3 is true
PASS axDiv3.name is "D"
PASS document.activeElement == div4 is true
PASS axDiv4.name is "Link"
PASS document.activeElement == div5 is true
PASS axDiv5.name is "Initial text before linkLink"
PASS document.activeElement == div6 is true
PASS axDiv6.name is "List item"
PASS document.activeElement == div7 is true
PASS axDiv7.name is "Initial text before list List item"
PASS successfullyParsed is true

TEST COMPLETE
